Day #148: WIP. folding+branching+abstraction+scaling. Code compiling, beginning to start runtime debugging.

+++ side_experiments/branch_fold_polish/solution/fold_steps/compress_scope_end_transform.cpp
@@ -21,7 +21,7 @@ void Fold::compress_scope_end() {
 		delete this->curr_fold;
 		this->curr_fold = this->test_fold;
 		this->test_fold = NULL;
-		for (int f_index = this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
+		for (int f_index = (int)this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
 			if (this->is_existing[f_index]) {
 				delete this->curr_input_folds[f_index];
 				this->curr_input_folds[f_index] = this->test_input_folds[f_index];
@@ -73,7 +73,7 @@ void Fold::compress_scope_end() {
 			this->test_fold = new FoldNetwork(this->curr_fold);
 			this->test_fold->pop_scope();
 			this->test_fold->add_scope(this->test_compressed_scope_sizes[0]);
-			for (int f_index = this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
+			for (int f_index = (int)this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
 				if (this->is_existing[f_index]) {
 					this->test_input_folds[f_index] = new FoldNetwork(this->curr_input_folds[f_index]);
 					this->test_input_folds[f_index]->pop_scope();
@@ -116,7 +116,7 @@ void Fold::compress_scope_end() {
 			this->test_fold->pop_scope();
 			this->test_fold->add_scope(this->test_compressed_scope_sizes[0]);
 			this->test_fold->add_scope(this->test_compress_new_size);
-			for (int f_index = this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
+			for (int f_index = (int)this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
 				if (this->is_existing[f_index]) {
 					this->test_input_folds[f_index] = new FoldNetwork(this->curr_input_folds[f_index]);
 					this->test_input_folds[f_index]->pop_scope();
@@ -145,7 +145,7 @@ void Fold::compress_scope_end() {
 
 		delete this->test_fold;
 		this->test_fold = NULL;
-		for (int f_index = this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
+		for (int f_index = (int)this->finished_steps.size()+1; f_index < this->sequence_length; f_index++) {
 			if (this->is_existing[f_index]) {
 				delete this->test_input_folds[f_index];
 				this->test_input_folds[f_index] = NULL;
